requirements.txt
aiogram>=3.15,<4.0
SQLAlchemy>=2.0,<3.0
asyncpg>=0.29
alembic>=1.13
python-dotenv>=1.0
redis>=5.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
testcontainers[postgresql]>=4.0.0
tzdata>=2024.1
